modificacion de mock de User y model del mismo
