+++ src/services/Users.ts
@@ -2,6 +2,7 @@ import axios from 'axios';
 import {API_URL} from '../constants';
 import users_mock from '../mock/Users/users.json';
 import user_mock from '../mock/Users/user.json';
+import usersByTerm from '../mock/Users/usersByTerm.json';
 
 const AxiosInstaceUser = axios.create({
   baseURL: API_URL,
@@ -22,7 +23,7 @@ class UserServiceMocked {
   static getUsersByTerm(query: string) {
     console.log('mock getUsersByTerm ', query);
     return Promise.resolve(() => {
-      return users_mock;
+      return usersByTerm;
     });
   }
   static getUser(username: string) {
@@ -61,60 +62,4 @@ export class UserService {
   }
 }
 
-export interface IUser {
-  login: string;
-  id: number;
-  node_id: string;
-  avatar_url: string;
-  gravatar_id: string;
-  url: string;
-  html_url: string;
-  followers_url: string;
-  following_url: string;
-  gists_url: string;
-  starred_url: string;
-  subscriptions_url: string;
-  organizations_url: string;
-  repos_url: string;
-  events_url: string;
-  received_events_url: string;
-  type: string;
-  user_view_type: string;
-  site_admin: boolean;
-}
 
-export interface IUserDetail {
-  login: string;
-  id: number;
-  node_id: string;
-  avatar_url: string;
-  gravatar_id: string;
-  url: string;
-  html_url: string;
-  followers_url: string;
-  following_url: string;
-  gists_url: string;
-  starred_url: string;
-  subscriptions_url: string;
-  organizations_url: string;
-  repos_url: string;
-  events_url: string;
-  received_events_url: string;
-  type: string;
-  user_view_type: string;
-  site_admin: boolean;
-  name: string;
-  company: string;
-  blog: string;
-  location: string;
-  email: any;
-  hireable: any;
-  bio: any;
-  twitter_username: string;
-  public_repos: number;
-  public_gists: number;
-  followers: number;
-  following: number;
-  created_at: string;
-  updated_at: string;
-}
